agentcore-runtime/parse-email-flight/requirements.txt
strands-agents>=0.1.0
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit
boto3>=1.34.0
requests>=2.31.0
pydantic>=2.0.0
